* refactors DatePicker component to use `useDatePicker` hook and splits logic into smaller utilities and `CalendarPopup` component
