+++ packages/ui-library/src/components/datePicker/index.ts
@@ -1,1 +1,2 @@
 export * from './datePicker';
+export * from './dateInput';
