+++ packages/ui-library/src/components/datePicker/types.ts
@@ -1,3 +1,4 @@
+import type {RefObject, KeyboardEvent} from 'react';
 import type {BaseProps} from '../common/types';
 
 export type DatePickerProps = BaseProps & {
@@ -26,3 +27,25 @@ export type CalendarDay = {
   currentMonth: boolean;
   disabled: boolean;
 }
+
+export type DatePickerPopupProps = {
+  ariaLabels: AriaLabels;
+  changeMonth: (offset: number) => void;
+  ID: string;
+  isNextDisabled: boolean;
+  isSameDay: (a: Date | null, b: Date | null) => boolean;
+  isPrevDisabled: boolean;
+  dayRefs: RefObject<Record<string, HTMLButtonElement | null>>;
+  focusedDate: Date | null;
+  handleDateChange: (date: Date) => void;
+  handleKeyDown: (e: KeyboardEvent) => void;
+  locale: string | undefined;
+  nextRef: RefObject<HTMLButtonElement | null>;
+  open: boolean;
+  prevRef: RefObject<HTMLButtonElement | null>;
+  today: Date;
+  value: Date | null;
+  view: Date;
+  weeks: CalendarDay[][];
+  weekdays: string[];
+};
